refactor: pull test config static class for reuse
